Move test class to conform convention
